Fix FileInfoExtensions.IsAssembly to work correctly with symbolic links.
The old behavior had an optimization of not bothering to read the file if it's smaller than 4K. This optimization doesn't work for symbolic links since the Length property of the FileInfo represents the length of the symbolic link which is always 0, rather than the file length of the target file.
The fix is to skip this optimization for symbolic links (identified by FileAttributes.ReparsePoint), and rely on the next check that attempts to actualy read the file and see if we got at least 4K of data.